Add line chart service
